Move item-related pretty printing functions to module
